Add :수정 페이지 생성 및 수정, 생성 페이지에 쓰이는 PostForm  컴포넌트화
