+++ src/styles/GlobalStyle.ts
@@ -1,9 +1,6 @@
 import { createGlobalStyle } from 'styled-components';
-import reset from 'styled-reset';
 
 const GlobalStyle = createGlobalStyle`
-${reset}
-
 * {
   box-sizing: border-box; 
   outline: none;
